refactor: move logic from HashAggregateExec into physical query planner
